requirements.txt
amqp==5.0.5
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.5
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.4.6
cssselect==1.1.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-admin==2.0.1
django-autoslug-iplweb==1.9.5
django-celery-beat==2.2.0
django-crispy-forms==1.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-nine==0.2.4
django-six==1.0.4
django-timezone-field==4.1.1
django-tinymce==3.2.0
djangorestframework==3.12.2
easy-thumbnails==2.7.1
feedparser==6.0.2
html5lib==1.0.1
hyperlink==21.0.0
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
ipython==7.20.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
kombu==5.0.2
lockfile==0.12.2
lxml==4.6.2
msgpack==0.6.2
ordereddict==1.1
packaging==20.3
parsel==1.6.0
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pif==0.8.2
Pillow==8.1.0
progress==1.5
prompt-toolkit==3.0.16
Protego==0.1.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
queuelib==1.5.0
requests==2.22.0
retrying==1.3.3
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
screen==1.0.1
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
transliterate==1.10.2
Twisted==20.3.0
Unidecode==1.2.0
urllib3==1.25.8
vine==5.0.0
vishap==0.1.5
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
xlwt==1.3.0
zope.interface==5.2.0
